Crate hsc

Simulate the dynamics of a stem cell population undergoing proliferation and differentiation according to a Moran process (fixed population size).

Whenever a cell divides, it acquires a Poisson number of passenger mutations, which do not give any proliferative advantage.

On top of that, cells can also acquire one proliferative advantageous mutation upon division. In this case, the cell creates a new clone, which has a birth-rate of lambda_i = lambda_0 ( 1 + s_i), where lambda_0 is the birth-rate of the wild-type, the clone without any proliferative mutations.

Modules

genotype

The neutral mutations representing the genotype of the stem cells.

process

The events to simulate for this Markov process.

proliferation

The proliferation of cells with the simulation of neutral and fit mutations.

stemcell

The agents whose state defines the system simulated by the process.

subclone

The classes defining the proliferative advantage.

Structs

ProbsPerYear

Enums

Probs

Probabilities used in the simulations.

Constants

MAX_SUBCLONES

Maximal number of fit clones that can arise during the simulation.

TIME_AT_BIRTH

The time at birth measured in years used for background mutations in the exponential growing phase.

Functions

write2file

Write vector of float into new file with a precision of 6 decimals. Write NAN if the slice to write to file is empty.