hsc

Module genotype

The neutral mutations representing the genotype of the stem cells.

Structs

MutationalBurden

Single-cell mutational burden is a mapping of cells sharing a number of mutations.

NeutralMutationPoisson

The Poisson probability distribution modeling the appearance of neutral mutations, which are assumed to follow a Poisson point process.

Sfs

Site frequency spectrum implemented as mapping with keys being j cells (x-axis) and values being the number of variants with j cells (y-axis).

Type Aliases

Variant