Trait hsc::proliferation::Proliferation

pub trait Proliferation {
    // Required method
    fn duplicate_cell_assign_mutations(
        &self,
        subclones: &mut SubClones,
        time: f32,
        proliferating_subclone: CloneId,
        distributions: &Distributions,
        rng: &mut impl Rng,
        verbosity: u8
    );
}

All the updates and changes in the system upon proliferation.

The proliferation step is implemented as following:

select the cell c that will proliferate next from the clone with id reaction determined by the Gillespie algorithm

draw and assign mb neutral background mutations to c from Poisson(DeltaT * mub), where mub is the backgound mutation rate

draw and assign md neutral division mutations to c from Poisson(mud), where mud is the division mutation rate

draw from a Bernoulli trial a fit mutation and in case assign c to a new clone. This clone must be empty and different from the old clone which c belonged to.

clone the proliferating cell c into c1 and repeat step 3, 4 with c1.

Required Methods

fn duplicate_cell_assign_mutations( &self, subclones: &mut SubClones, time: f32, proliferating_subclone: CloneId, distributions: &Distributions, rng: &mut impl Rng, verbosity: u8 )

Object Safety

This trait is not object safe.