hsc::stemcell

Function assign_background_mutations

pub fn assign_background_mutations(
    stem_cell: &mut StemCell,
    time: f32,
    neutral_poisson: &NeutralMutationPoisson,
    rng: &mut impl Rng,
    verbosity: u8,
)

Assign background mutations to all cells in the system based on the current simulation time.

This updates also the time of the last division for the cell.