hsc

Module process

The events to simulate for this Markov process.

Structs

CellDivisionProbabilities

Exponential

Exponential growing process

Moran

The Moran process saves the state of the agents and simulates new proliferative events at each timestep according to the Gillespie algorithm, see Moran::advance_step.

ProcessOptions

SavingOptions

Snapshot

Enums

ExponentialPhase

SavingCells

Stats2Save

Functions

switch_to_moran

End the exponential growing phase and switch to a fixed-size population phase.