hsc::process

Function switch_to_moran

pub fn switch_to_moran(
    exponential: Exponential,
    process_options: ProcessOptions,
    distributions: Distributions,
    filename: PathBuf,
    save_sfs_only: bool,
    save_population: bool,
    rng: &mut impl Rng,
) -> Moran

End the exponential growing phase and switch to a fixed-size population phase.

There is a delay between birth and the end of the exponential phase, since stem cells stop dividing exponentially before birth. We add background mutations in this interval of time between the end of the exponentially growing phase and the Moran process.