hsc

Module subclone

The classes defining the proliferative advantage.

Structs

Distributions

Distribution probabilities for the simulations upon cell division.

SubClone

A group of cells sharing the same genetic background with a specific proliferation rate.

SubClones

A collection of subclones each having their proliferative advantage. This collection contains the neutral clone as well.

Variants

Number of cells in subclones.

Enums

Fitness

Fitness models implemented so far.

Functions

from_mean_std_to_shape_scale

next_clone

Returns a CloneId to which the stem cell must be assigned by sampling a fit variant.

proliferating_cell

Determine which cells will proliferate by randomly selecting a cell from the subclone with id subclone_id. This will also remove the random selected cell from the subclone, hence subclones are borrowed mutably.

save_variant_fraction

Type Aliases

CloneId

Id of the SubClones.