Function hsc::subclone::proliferating_cell

pub fn proliferating_cell(
    subclones: &mut SubClones,
    subclone_id: CloneId,
    verbosity: u8,
    rng: &mut impl Rng
) -> StemCell

Determine which cells will proliferate by randomly selecting a cell from the subclone with id subclone_id. This will also remove the random selected cell from the subclone, hence subclones are borrowed mutably.